loleaflet: Resurrect loleaflet permission model for readonly mode
Remove direct usage of editlock; instead change loleaflet
permission from view to edit when editlock message is received.
By default, map would be in 'view' mode and only get to 'edit'
when server asks it to switch to.

Use a new event 'doclayerinit' for actions that are directly
dependent on intialization of doclayer.

Change-Id: Iaceb95ad85721ddcd675c75cfeb1504ab2759ad4